ui: display items collected text
