+++ BlasII.Randomizer/SystemPatches.cs
@@ -1,9 +1,7 @@
 ﻿using HarmonyLib;
 using Il2CppTGK.Game.Components.Misc;
-using Il2CppTGK.Game.Components.UI;
 using Il2CppTGK.Game.DialogSystem;
 using Il2CppTGK.Game.Managers;
-using Il2CppTGK.Game.PopupMessages;
 using System.Reflection;
 
 namespace BlasII.Randomizer
@@ -48,34 +46,6 @@ namespace BlasII.Randomizer
         public static void Prefix(Dialog dialog)
         {
             Main.Randomizer.Log("Starting dialog: " + dialog.name);
-        }
-    }
-
-    /// <summary>
-    /// When reading the CR door, show how many keys you must find
-    /// When pressing display button, show the current settings
-    /// </summary>
-    [HarmonyPatch(typeof(PopupMessageLogic), nameof(PopupMessageLogic.ShowMessageAndWait))]
-    class Popup_Show_Patch
-    {
-        public static void Postfix(PopupMessageLogic __instance, PopupMessage message)
-        {
-            Main.Randomizer.Log("Showing popup: " + message.name);
-
-            if (message.name == "MSG_0003")
-            {
-                string text = Main.Randomizer.LocalizationHandler.Localize("ktex")
-                    .Replace("*", Main.Randomizer.CurrentSettings.RealRequiredKeys.ToString());
-                __instance.textCtrl.SetText(text);
-                return;
-            }
-
-            if (message.name == "TESTPOPUP")
-            {
-                string text = Main.Randomizer.CurrentSettings.FormatInfo();
-                __instance.textCtrl.SetText(text);
-                return;
-            }
         }
     }
 }
